release: oauth, sdks, docs, ui cleanup

+++ others/python-sdk/src/autumn_sdk/models/tracktokensop.py
@@ -63,7 +63,7 @@ class TrackTokensParamsTypedDict(TypedDict):
     r"""Number of reasoning tokens generated."""
     properties: NotRequired[Dict[str, Any]]
     r"""Additional properties to attach to this usage event."""
-    timestamp: NotRequired[float]
+    timestamp: NotRequired[int]
     r"""Unix timestamp in milliseconds to use for the usage event. Defaults to the current time."""
     async_: NotRequired[bool]
     r"""If true, enqueue the event for asynchronous processing and return 204 immediately. The response will not include balance information."""
@@ -106,7 +106,7 @@ class TrackTokensParams(BaseModel):
     properties: Optional[Dict[str, Any]] = None
     r"""Additional properties to attach to this usage event."""
 
-    timestamp: Optional[float] = None
+    timestamp: Optional[int] = None
     r"""Unix timestamp in milliseconds to use for the usage event. Defaults to the current time."""
 
     async_: Annotated[Optional[bool], pydantic.Field(alias="async")] = None
